requirements.txt
dash==2.14.2
datasets==2.16.0
gensim==4.2.0
matplotlib==3.7.4
nltk==3.8.1
numpy==1.23.0
octis==1.13.1
pandas==2.0.3
plotly==5.18.0
pyarrow==14.0.2
scikit_learn==1.1.0
scipy==1.9.3
seaborn==0.13.1
sentence_transformers==2.2.2
setfit==1.0.1
setuptools==68.2.2
tqdm==4.65.0
umap_learn==0.5.4
wordcloud==1.9.3
yellowbrick==1.5
